LEAN-1190 - LW -R5: No successful message appeared after uploading file to dashboard inspector
